Refactor worker job
