+++ include/simplemapreduce/ops/conf.h
@@ -11,7 +11,7 @@ namespace mapreduce {
   struct JobConf {
     /* # of groups to split tasks */ int n_groups{10};
     /* Temporary directory path */   std::filesystem::path tmpdir{"./_tmp"};
-    /* Output file path */           std::filesystem::path output_fpath;
+    /* Output file path */           std::filesystem::path output_dirpath;
     /* # of nodes */                 int mpi_size;
     /* Current rank in mpi */        int mpi_rank;
     /* # of worker to run tasks */   int worker_size;
