+++ src/job.cc
@@ -9,6 +9,7 @@
 #include <type_traits>
 
 #include "simplemapreduce/data/bytes.h"
+#include "simplemapreduce/local/runner.h"
 #include "simplemapreduce/proc/writer.h"
 #include "simplemapreduce/util/argparse.h"
 #include "simplemapreduce/util/log.h"
@@ -16,6 +17,7 @@
 
 namespace fs = std::filesystem;
 
+using namespace mapreduce::base;
 using namespace mapreduce::commons;
 using namespace mapreduce::data;
 using namespace mapreduce::proc;
@@ -54,6 +56,9 @@ Job::Job(int &argc, char *argv[]) {
 }
 
 void Job::start_up() {
+  runner_ = std::make_unique<mapreduce::local::LocalJobRunner>();
+  runner_->set_conf(conf_);
+
   /// Initialize connection and get the status
   MPI_Comm_rank(MPI_COMM_WORLD, &(conf_->mpi_rank));
   MPI_Comm_size(MPI_COMM_WORLD, &(conf_->mpi_size));
@@ -113,15 +118,6 @@ void Job::setup_output_dir() {
   }
 }
 
-fs::path Job::get_output_file_path() {
-  /// Set file path named by current worker rank and join with the output directory
-  std::ostringstream oss;
-  oss << std::setw(5) << std::setfill('0') << conf_->worker_rank;
-  fs::path fname = oss.str();
-  
-  return file_fmt_.get_output_path() / fname;
-}
-
 /* --------------------------------------------------
  *   Helper functions
  * -------------------------------------------------- */
@@ -137,22 +133,6 @@ int Job::find_available_worker() {
   int index;
   MPI_Waitany(mpi_reqs.size(), mpi_reqs.data(), &index, mpi_worker_statuses.data());
   return index;
-}
-
-std::string Job::receive_filepath() {
-  MPI_Status status;
-  MPI_Probe(0, TaskType::map_data, MPI_COMM_WORLD, &status);
-
-  /// Get data size to reveive for receiving file path.
-  /// Lengths of file path are vary so that need to check data size first.
-  int data_size;
-  MPI_Get_count(&status, MPI_CHAR, &data_size);
-
-  /// Receive target file path to apply map task
-  char path_data[data_size];
-  MPI_Recv(path_data, data_size, MPI_CHAR, status.MPI_SOURCE, status.MPI_TAG, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
-
-  return std::string(path_data, data_size);
 }
 
 /* --------------------------------------------------
@@ -228,108 +208,15 @@ void Job::start_master_node() {
 }
 
 /* --------------------------------------------------
- *   Child Node
+ *   Main
  * -------------------------------------------------- */
-void Job::run_child_tasks() {
-  conf_->output_fpath = get_output_file_path();
-
-  /// Run mapreduce processes
-  run_map_tasks();
-  run_shuffle_tasks();
-
-  /// Send signal to notify finished tasks up to shuffle
-  MPI_Send("\0", 1, MPI_CHAR, 0, TaskType::shuffle_end, MPI_COMM_WORLD);
-
-  /// Receive signal to resume tasks
-  char tmp;
-  MPI_Recv(&tmp, 1, MPI_CHAR, 0, TaskType::sort_start, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
-
-  /// Resume after all processes up to shuffle in all nodes has been finished
-  run_reduce_tasks();
-
-  /// Send signal to notify finished all tasks
-  MPI_Send("\0", 1, MPI_CHAR, 0, TaskType::reduce_end, MPI_COMM_WORLD);
-
-  logger.debug("[Worker] Finished all tasks on worker ", conf_->worker_rank);
-}
-
-void Job::run_map_tasks() {
-  /// Send signal to notify enqueue step is finished
-  auto mq = mapper_->get_mq();
-
-  std::future<void> combiner_ftr;
-
-  if (combiner_ != nullptr) {
-    /// Start Combiner
-    combiner_->set_mq(mq);
-    combiner_ftr = std::async(std::launch::async, [&]{ combiner_->run(); });
-  }
-
-  while (true) {
-    char req;
-    MPI_Recv(&req, 1, MPI_CHAR, 0, TaskType::ready, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
-
-    /// Once received end of process signal, break the loop
-    if (req == '\0') 
-      break;
-
-    /// Check the future status if already run any task
-    /// Send to notify this worker is available
-    MPI_Send("\1", 1, MPI_CHAR, 0, TaskType::map_start, MPI_COMM_WORLD);
-
-    /// Once received empty string, stop receiving data
-    std::string target_path = receive_filepath();
-
-    logger.debug("[Worker] Assigned a file: \"", target_path, "\" to worker ", conf_->worker_rank);
-
-    /// Read data from text file
-    {
-      std::ifstream ifs(target_path);
-      std::ostringstream oss;
-      oss << ifs.rdbuf();
-      ifs.close();
-
-      /// Read data from text file
-      ByteData key{oss.str()}, value{1l};
-      oss.clear();
-
-      /// Start map task
-      mapper_->run(key, value);
-    }
-
-    /// Notify the map task is finished
-    MPI_Send("\1", 1, MPI_CHAR, 0, TaskType::map_end, MPI_COMM_WORLD);
-  }
-
-  /// Send signal to the end of Map
-  mq->end();
-
-  if(combiner_ != nullptr) {
-    logger.debug("[Worker] Running Combiner on worker ", conf_->worker_rank);
-    /// Wait until Combiner end and send signal to notify the end of the process
-    combiner_ftr.get();
-    mq->end();
-  }
-
-  logger.debug("[Worker] Finished Map on worker ", conf_->worker_rank);
-}
-
-void Job::run_shuffle_tasks() {
-  auto shuffler = mapper_->get_shuffle();
-  shuffler->run();
-}
-
-void Job::run_reduce_tasks() {
-  reducer_->run();
-}
-
 int Job::run() {
   /// Not run when -h/--help option is passed
   if (!is_valid_) return 0;
 
   /// send target file to each node
   if (conf_->mpi_rank == 0) {
-    if (mapper_ == nullptr || reducer_ == nullptr)
+    if (!has_mapper_ || !has_reducer_)
       throw std::runtime_error("Mapper and/or Reducer is not set.");
 
     /// Set up a directory to store intermediate state data
@@ -341,7 +228,9 @@ int Job::run() {
     fs::remove_all(conf_->tmpdir);
     logger.info("[Master] Finished task");
   } else {
-    run_child_tasks();
+    conf_->output_dirpath = file_fmt_.get_output_path();
+    runner_->start();
+    logger.debug("[Worker] Finished all tasks on worker ", conf_->worker_rank);
   }
 
   return 0;
